# legacy-codebase/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MyDreamCampus API
  description: |
    Üniversite yönetim sistemi API dokümantasyonu.

    Tüm API'ler Traefik gateway üzerinden `http://localhost/api/...` şeklinde erişilir.

    ## Kimlik Doğrulama
    - `POST /api/auth/login` ile JWT token alınır
    - Token `Authorization: Bearer <token>` header'ı ile gönderilir
    - Traefik forward-auth middleware'i token'ı doğrular ve `X-User-ID`, `X-User-Role`, `X-User-Department` header'larını backend servislere iletir
    - Refresh token HttpOnly cookie olarak saklanır

    ## Roller
    - **admin**: Tam yetki
    - **teacher**: Ders, yoklama, not yönetimi
    - **student**: Kayıt, yoklama, not görüntüleme

    ## Rate Limiting
    Tüm endpoint'ler rate limiting ile korunur. Yanıt header'ları:
    - `X-RateLimit-Limit`: Zaman penceresi başına izin verilen maksimum istek sayısı
    - `X-RateLimit-Remaining`: Kalan istek sayısı
    - `X-RateLimit-Reset`: Limit sıfırlanma zamanı (Unix epoch saniye)

    Limit aşıldığında `429 Too Many Requests` döner.

    ## Hata Yönetimi
    Tüm endpoint'ler aşağıdaki genel hata yanıtlarını dönebilir:
    - `401 Unauthorized`: Geçersiz veya eksik JWT token
    - `429 Too Many Requests`: Rate limit aşıldı
    - `500 Internal Server Error`: Beklenmeyen sunucu hatası

    Hata yanıtları her zaman `ErrorResponse` formatındadır.

    ## Sağlık Kontrolü
    Her servis `GET /health` endpoint'i sunar. Bu endpoint'ler kimlik doğrulama gerektirmez.
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/semesters/{semester_id}/courses:
    get:
      operationId: listSemesterCourses
      tags: [Semesters]
      summary: Dönem derslerini listele
      parameters:
        - name: semester_id
          in: path
          required: true
          schema:
            type: string
          example: "2025-2026-spring"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: faculty
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: instructor_id
          in: query
          schema:
            type: string
            format: uuid
        - name: course_type
          in: query
          schema:
            type: string
            enum: [mandatory, elective]
        - name: class_level
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 6
      responses:
        "200":
          description: Dönem ders listesi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemesterCourseListResponse"

    post:
      operationId: createSemesterCourse
      tags: [Semesters]
      summary: Döneme ders ekle
      description: Sadece admin.
      parameters:
        - name: semester_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSemesterCourseRequest"
      responses:
        "201":
          description: Dönem dersi oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemesterCourseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/semesters/{semester_id}/courses/{course_id}:
    get:
      operationId: getSemesterCourse
      tags: [Semesters]
      summary: Dönem ders detayı
      parameters:
        - name: semester_id
          in: path
          required: true
          schema:
            type: string
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dönem ders bilgisi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemesterCourseResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteSemesterCourse
      tags: [Semesters]
      summary: Dönemden ders sil
      description: Sadece admin.
      parameters:
        - name: semester_id
          in: path
          required: true
          schema:
            type: string
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dönem dersi silindi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSemesterCourseResponse"
  /api/semesters/teacher/courses:
    get:
      operationId: getTeacherCourses
      tags: [Semesters]
      summary: Öğretmenin derslerini getir
      description: teacher rolü gerektirir.
      parameters:
        - name: semester
          in: query
          schema:
            type: string
          description: Belirtilmezse aktif dönem kullanılır
      responses:
        "200":
          description: Öğretmen ders listesi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherCoursesResponse"

  # ============================================================
  # ENROLLMENT SERVICE
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. `POST /api/auth/login` ile alınır."
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Geçersiz istek
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Kayıt bulunamadı
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "email alanı gereklidir"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8

    # ---- Auth ----
...
    Prerequisite:
      type: object
      properties:
        id:
          type: string
          format: uuid
        course_code:
          type: string
        course_name:
          type: string
...
    ScheduleSession:
      type: object
      required: [day_of_week, slot_numbers, session_type]
      properties:
        day_of_week:
          type: string
          enum: [monday, tuesday, wednesday, thursday, friday]
          example: "monday"
        slot_numbers:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 12
          example: [1, 2]
        session_type:
          type: string
          enum: [theory, lab]
          example: "theory"
      example:
        day_of_week: "monday"
        slot_numbers: [1, 2]
        session_type: "theory"
    AssessmentItem:
      type: object
      required: [slug, name, weight]
      properties:
        slug:
          type: string
          example: "midterm"
          description: Değerlendirme tanımlayıcısı (benzersiz, slug formatında)
        name:
          type: string
          example: "Ara Sınav"
        weight:
          type: integer
          minimum: 1
          maximum: 100
          description: Yüzde ağırlık (toplam 100 olmalı)
          example: 40
      example:
        slug: "midterm"
        name: "Ara Sınav"
        weight: 40
    CreateSemesterCourseRequest:
      type: object
      required:
        - course_code
        - class_level
        - instructor_id
        - instructor_fullname
        - classroom_location
        - max_capacity
        - assessment_schema
        - schedule_sessions
      properties:
        course_code:
          type: string
        class_level:
          type: integer
          minimum: 1
          maximum: 6
        instructor_id:
          type: string
          format: uuid
        instructor_fullname:
          type: string
        classroom_location:
          type: string
        max_capacity:
          type: integer
          minimum: 1
          maximum: 1000
        assessment_schema:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AssessmentItem"
        schedule_sessions:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ScheduleSession"
    SemesterCourseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        semester:
          type: string
        course_code:
          type: string
        course_name:
          type: string
        credits:
          type: integer
        class_level:
          type: integer
        instructor_id:
          type: string
          format: uuid
        instructor_fullname:
          type: string
        classroom_location:
          type: string
        max_capacity:
          type: integer
        assessment_schema:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentItem"
        schedule_sessions:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleSession"
        prerequisites:
          type: array
          items:
            $ref: "#/components/schemas/Prerequisite"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SemesterCourseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SemesterCourseResponse"
        pagination:
          $ref: "#/components/schemas/Pagination"
    DeleteSemesterCourseResponse:
      type: object
      properties:
        message:
          type: string
        semester_course_id:
          type: string
        course_code:
          type: string
        semester:
          type: string
    TeacherCoursesResponse:
      type: object
      properties:
        instructor_id:
          type: string
          format: uuid
        total_courses:
          type: integer
        courses:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              course_code:
                type: string
              course_name:
                type: string
              faculty:
                type: string
              department:
                type: string
              semester:
                type: string
              credits:
                type: integer
              theoretical_hours:
                type: integer
              lab_hours:
                type: integer
              classroom_location:
                type: string
              max_capacity:
                type: integer
              schedule:
                type: array
                items:
                  type: object
                  properties:
                    day:
                      type: string
                    time:
                      type: string
                    room:
                      type: string
                    session_type:
                      type: string

    # ---- Enrollment ----
...